refactor: move add-plugin logic and ui in plugin-list-view
